refactor(cli): extract runtime command module (status, dispatch)
